Hide window borders along with UI
I'm using the COSMIC desktop environment.  When I capture the
SnekStudio window (using the ScreenCast portal), it includes the
compositor-rendered window decorations.  This patch makes it so that
when the UI is hidden, so are the window decorations, so that it's
possible to capture just the contents of the window, even with capture
implementations that include the decorations.