Import textures in S3TC and ASTC
It seems Godot by default imports textures in the format optimal for the
host platform, then gets very confused if it decides it needs textures
in another format for some reason (such as project export settings),
and re-imports forever.

Force texture import in both main formats (ASTC, S3TC). For
cross-platform collaboration, this should make sure there's no weirdness
and the .import files are also universal. It's a bit slower, but pretty
trivial for a projec this size.