Core/UI: Fix window moving down when popping in
When popping a window back in after it's been popped out, the reparent()
call will cause the window to reposition relative to its parent, which
causes the embedded version of a window to move down a bit, so if you
don't want the position of a window to be modified, you have to set the
window position/size *after* the reparent() call. Learned this after
debugging into the engine code itself.