Add SNEKSTUDIO_SAMPLE_PATH environment variable
Prior to this commit snekstudio needed to be launched in the same
directory as where the SampleModels folder is present. This isn't a
ideal. To solve this, this commit introduces SNEKSTUDIO_SAMPLE_PATH and
get_sample_location. if SNEKSTUDIO_SAMPLE_PATH is set,
get_sample_location returns the localized path, if it is not set it
returns res://SampleModels/VRM/ effectively preserving the behavior of
snekstudio with out it being set.