Make CI build flatpaks
This adds a flatpak as a build artifact. The flatpaks are built from the
tarballs produced by the build script. This however requires 2 jobs as
flatpak can not install the required Sdk and Platform inside a container
as it requires a running dbus impl.

To work around this, this break the existing workflow into 2 jobs. First
we build snudio as normal, then instead of uploadingit as a release,
we then upload them as build artifacts. Then the second job downloads
the built artifacts and builds the flatpaks before publishing a release.

We can greatly reduce the build time for the flatpak if we have a icon
for the flatpak already scaled to the following sizes, 32x32, 64x64,
128x128, 256x256, 512x512. This would mean we can drop building
imagemagick and just bundle snudio.